Moves solutions into separate folder
