+++ solutions/optimal.js
